+++ sg/app/variables.tf
@@ -107,9 +107,16 @@ variable "vpc_key" {
   type = string
 }
 
-variable "vpc_map" {
+variable "vpc_data_map" {
+  # Add attributes as needed for a drop-in replacement
   type = map(object({
-    name_simple         = string
+    name_simple           = string
+    security_group_id_map = map(string)
+    segment_map = map(object({
+      route_public  = bool
+      subnet_id_map = map(string)
+    }))
+    vpc_cidr_block      = string
     vpc_id              = string
     vpc_ipv6_cidr_block = string
   }))